requirements.txt
Django==4.2.5
Pillow==8.0.1
asgiref==3.6.0
pytz==2021.1
sqlparse==0.4.1
django-cors-headers==3.8.0
django_custom_user==1.1
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
django-rest-swagger==2.2.0
django-extra-fields==3.0.2
django-filter==2.4.0
drf-yasg==1.21.7
PyJWT==1.7.1
ipython==7.16.1
psycopg2-binary==2.9.1
psycopg2==2.9.1
sentry-sdk==0.20.0
coreapi==2.3.3
coreschema==0.0.4
django-storages==1.10.1
MarkupSafe==1.1.1
celery==5.0.5
django-celery-beat==2.1.0
django-celery-results==2.0.0
sorl-thumbnail==12.7.0
redis==4.2.0
django-redis-sessions==0.6.1
django-redis==4.12.1
environ==1.0
python-dotenv==0.17.0
gunicorn==20.1.0
uvicorn==0.13.4
uvloop==0.14.0
httptools==0.1.1
hashids==1.3.1
google-cloud-storage==1.38.0
oauth2client==4.1.3
django-countries==7.1
django-mathfilters==1.0.0
clicksend-client==5.0.72
google-api-python-client==2.2.0
django-stdimage==5.3.0
google-auth-oauthlib==0.4.4
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0
xlwt==1.3.0
openpyxl==3.0.9
pdf2image==1.16.0
requests==2.23.0
python-dateutil==2.8.2
ezdxf==0.17.2
jsonschema==4.16.0
whitenoise==6.2.0
numpy==1.23.3
scipy==1.9.1
weasyprint==57.0
flower==1.2.0
django-prometheus==2.3.1
dj-database-url==1.2.0
